Changed CTcal2 to save resulst from new energy charge-trapping correction algorithms. Added sortskim2 and sortrun2 to properly handle new energy charge-trapping correction algorithms. Fixed bug in sortrun that pevented interpolation of A/E cut between calibrations.

+++ mjd/merge_skim.c
@@ -47,7 +47,7 @@ int main(int argc, char **argv) {
     if ((c = strstr(ds_fname, "/skim.dat"))) {
       *c = 0;
     } else {
-      printf("\n ERROR: No directory found in %s; c = %d\n", ds_fname, (int) c);
+      printf("\n ERROR: No directory found in %s; c = %lld\n", ds_fname, (long long int) c);
       exit(-1);
     }
     sprintf(CTC.ctc_fname, "%s/ctc.input", ds_fname);
